added functionality to fill the pdf forms from a given config cleanup
