+++ PDF Form Filler/Config.cs
@@ -14,24 +14,8 @@ namespace PDF_Form_Filler
 
         public void Load(string path)
         {
-            //var config = new Config();
-
-            //// Load and deserialize settings
-            //if (!File.Exists(path))
-            //    throw new Exception($"config not found at {Directory.GetCurrentDirectory()}");
-
-            //string json = File.ReadAllText(path);
-            //config = JsonSerializer.Deserialize<Config>(json);
-
-            //return config;
-            
-            
-            
-            
-            
             var config = new Config();
 
-            // Load and deserialize settings
             if (!File.Exists(path))
                 throw new Exception($"config not found at {Directory.GetCurrentDirectory()}");
 
@@ -42,20 +26,12 @@ namespace PDF_Form_Filler
 
         public void Save(string filePath)
         {
-            // Serialize the Config object to JSON
             string json = JsonSerializer.Serialize(this, new JsonSerializerOptions
-            {
-                WriteIndented = true // Makes the JSON output more readable with indentation
-            });
-
-            // Write the JSON to a file
+                {
+                    WriteIndented = true
+                }
+            );
             File.WriteAllText(filePath, json);
         }
-
-        //public List<KeyValuePair<string, string>> Kunde { get; set; }
-        //public List<KeyValuePair<string, string>> Selbst { get; set; }
-        //public List<KeyValuePair<string, string>> Textfields { get; set; }
-        //public List<KeyValuePair<string, string>> Checkboxes { get; set; }
-        //public List<KeyValuePair<string, string>> Groups { get; set; }
     }
 }
