+++ PDF Form Filler/Program.cs
@@ -16,7 +16,7 @@ namespace PDF_Form_Filler
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new MainForm());
         }
     }
 }
